[OWL-1332][nqm-mng][common] test code refinement
